chore(workaround): getting esri.map module is not defined error during dgeni build
angular.module(esri.map,[])... was embedded in esriLoader.js when dgeni is processing esriFeatureLayer.js
first, it did not see the esri.map module, and throws the above mentioned error. The workaround was to move the
module definition outside of the esriLoader.js and create a separate set of files to be processed first (see
dgeni-fgpv.js). This might not being an issue with the dgeni-conf.js